+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Restart background thread when FD's are closed
 * Beta version of collecting host metrics (disabled by default)
 * Hooks for Shuryoken
+* Don't add errors from env if raise_errors is off for Sinatra
 
 # 1.1.9
 * Fix for race condition when creating working dir exactly at the same time